check if student can graduate
Added a student find field to the student section of the program. Right now when a student types their student ID into the textbox and clicks find it checks to see if they have enough credits to graduate. This will eventually (or that's the plan) load in the student information for use on each of the student sections.